Combined Scripts of Nebula and NebulaGel -> Nebula
- combined codes from Nebula and NebulaGel into one script under name
Nebula
- damaging nebula is now set to type 3
- code in previous Nebula remains unchanged, exists within type 3
condition
