--- Assets/Scripts/NebulaGel.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 25eaee219c3c54d4a960130225300fe7
-timeCreated: 1474255407
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
